ci: Move to full prebuilt buildroot images
Building the images can take too long on Travis and it's not like we
need bleeding edge images either. The cpio image is for booting with
the '-initrd' flag, as opposed to the '-hda' flag + 'root=' on the
cmdline.

While we're at it, switch from 'reboot' in the init.d file and
'-no-reboot' on the QEMU cmdline to just 'poweroff' in the init.d file,
which makes it a little clearer that's what we intend to do.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>