ci: Use ccache
Installing via Travis allows it to properly set up all of the
directories and paths but we need to install 3.5 so that we don't
see a ton of warnings from included and preprocessed code that we
shouldn't see. This was fixed in 3.2 and Trusty's version is 3.1.9.

https://ccache.samba.org/releasenotes.html#_ccache_3_2

Reading the release notes, we also get some nice Clang improvements as
well and it only takes a few seconds to build.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>