driver.sh: Use KBUILD_BUILD variables to make builds more reproducible
As Nick pointed out at the issue below, we might not getting the most
out of ccache because some environment variables are dynamic (date and
user/host on Travis). Provide static values for these so we actually
see speedup with ccache.

Locally, I've verified this works with an arm64 defconfig build.

Cold cache: 10m57.146s
Hot cache: 0m42.958s

Unfortunately, it doesn't seem like this speeds up Travis for arm64 but
it does for arm32 and x86_64.

Closes #20.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>