ci: Switch to Docker
Unfortunately, as we add new architectures, we are going to need more up
to date binaries that Trusty cannot provide (see #16). Move our building
infrastructure to Docker, which gives us a deterministic, configurable
environment every time. The only forseeable downside is no cacheable
ccache directory so slower builds but we are getting to a point where
constant development is slowing down. I have kept ccache support in
driver.sh so that anyone who uses it locally can still take advantage of
it.

Initially, we'll start with the latest Debian stable but if we need to
move to the unstable version (sid) for more up to date packages, it's as
easy as changing 'debian' to 'debian:sid' in .travis.yml and adjusting
the LLVM/Clang apt repo in debian-setup.sh.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>