.travis.yml: Add linux-next to the daily cron
We don't need to build this on every branch push or pull request, it
only changes once a day and there is nothing specific to it that needs
regular testing (torvalds/linux will suffice).

Add a comment above each block so that it is easy to see which block
corresponds to which repo.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>