debian-setup: quiet down setup, and zero stats
The logs are full of noise about packages being installed. This is
noise.

Printing the stats before and after makes it hard to tell the difference
without doing math.  It's clearer to zero stats before. Then the stats
after are just for this build.  We don't care about the cache stats pre
run.  Clearing the stats DOES NOT clear the cache itself or reset its
size.

https://ccache.samba.org/manual/latest.html#_options