patches: x86_64: Remove Makefile change
I had this in my personal patchset because according to https://github.com/ClangBuiltLinux/linux/issues/14,
the CONFIG_ORC_UNWINDER check was causing the build to error. I never
actually verified that result myself. Along with #48 and this patch,
there are no warnings/errors[1]. The biggest reason for removing the
changes to this file are to avoid potential merge conflicts[2].

[1]: https://travis-ci.com/nathanchance/continuous-integration/builds/91060888
[2]: https://lore.kernel.org/lkml/1539681053-24388-3-git-send-email-yamada.masahiro@socionext.com/

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>