driver: Only make dtbs on arm targets
Commit 59cf97d0bdab ("driver: build dtbs target") builds the dtbs for
all targets except x86_64. However, this causes issues with PowerPC on
LTS branches prior to 4.20 due to lack of commit 37c8a5fafa3b ("kbuild:
consolidate Devicetree dtb build rules") in the kernel tree because
PowerPC did not have a dtbs target. This change is too invasive for an
LTS backport in my opinion and doesn't really fix a bug anyways. We
currently do not use any dtbs on PowerPC; if this changes in the future,
we can come up with a more clever way of working around this.

Fixes the following Travis failures:

https://travis-ci.com/ClangBuiltLinux/continuous-integration/jobs/167579254

https://travis-ci.com/ClangBuiltLinux/continuous-integration/jobs/167579258

[skip ci]

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>