driver.sh: Add support for kernel/common
kernel/common is important as it is the base kernel for all Android
devices so we should make sure it never regresses with Clang.

A couple of implementation details:

* URL needs to be overridable because this tree is not on kernel.org.

* cuttlefish's x86_64 target needs to be booted to a ramdisk as it does
  not support '/dev/sda'.

It would be nice to test LTO and CFI to avoid issues like
https://github.com/ClangBuiltLinux/linux/issues/406 but there is a zero
percent chance that the VM could handle that (for both memory
and time reasons). I will test these locally.

Some TODOs (I can file these as other issues if needed):

* Inquire as to why CONFIG_FRAME_WARN is set to 1024, when the default
  is 2048 for 64-bit. This introduces a few warnings.

* Investigate other warnings and backport the necessary fixes to stable.

* Maybe run some other tests while booted like the VTS kernel suite?

Closes: #91
Suggested-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>