driver: Improvements around binary locating
We've tailored driver.sh around our Docker image with regards to the
binary versions that we check for. A user might not have the latest
versions of Clang or llvm-ar installed. Check for all three versions
offered by apt.llvm.org, falling back to the generic binary name if
those are not found.

Additionally, for llvm-ar, do a sanity check that the user has a version
of llvm-ar that can actually build the kernel (has support for the P
modifier). If they don't fall back to regular GNU ar and let them know.

Presubmit: https://travis-ci.com/nathanchance/continuous-integration/builds/103470578

[skip ci]

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>