ppc64le: Add skiboot image
The Debian skiboot build fails to reboot or shutdown with Qemu 3.1. This
has been fixed in skiboot master but not yet in the Debian package.

The image added to the tree comes from the OpenPower CI system:

 https://openpower.xyz/job/openpower/job/openpower-op-build/label=slave,target=palmetto/

It is built from https://github.com/open-power/skiboot.

This change should be reverted when Debian releases a fixed package.

Resolves https://github.com/ClangBuiltLinux/continuous-integration/issues/137
Signed-off-by: Joel Stanley <joel@jms.id.au>