driver: Disable LTO/CFI by default
This is now enabled by default on Cuttlefish 4.19, which causes issues
with our Debian setup because we do not install the regular LLVM
packages so llvm-nm is not available (llvm-nm-9 is). There is currently
no way to override LLVM_NM's assignment in Makefile:

https://github.com/aosp-mirror/kernel_common/blob/50f91435a27a9f7/Makefile#L617

It might be worth making that customizable in the future by changing the
:= into ?=; however, this is a moot point for our CI setup because
Travis cannot handle LTO for both time and memory reasons so just
disable it by default. A user can still pass '--lto' to driver.sh
to avoid disabling LTO (it will NOT attempt to enable it).

Fixes: https://travis-ci.com/ClangBuiltLinux/continuous-integration/jobs/202781149

Presubmit: https://travis-ci.com/nathanchance/continuous-integration/builds/113113278

[skip ci]

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>