ci: Don't apply asm goto hacks when using LLVM 9 and above
We want to test LLVM 8 in CI because it is the current release version
and we want to clearly see if there is a kernel regression or LLVM
regression. If LLVM 8 is fine but LLVM 9 causes errors, it's probably
something on the LLVM side rather than kernel.

The folder structure will now be:

llvm-all / llvm-version ->
    kernel-all / ${REPO} ->
        arch-all / ${ARCH}

Hopefully we never have to use the all folders :)

I have hesistantly added symlinks to llvm-6 and llvm-7 on the off chance
that others want to test older versions (for example, Ubuntu Bionic, the
latest LTS, has clang-6 in its repos), even though they really should be
using a newer version.

Presubmit: https://travis-ci.com/nathanchance/continuous-integration/builds/114532215

Closes: #167
Closes: #171
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>