driver.sh: Generate LLVM version arrays dynamically
We currently check for versions of five different LLVM binaries,
resulting in a lot of duplicated code, which all needs to be updated
when an LLVM version bump occurs.

To simply this, we will use seq and its '-f' parameter to dynamically
create this array based on the name of the binary and the latest/oldest
LLVM version we support.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>