driver: Support booting via EFI on arm64 and x86_64
If the ovmf or qemu-efi-aarch64 packages are installed on Debian or
Ubuntu, use those files to boot up via EFI. This will help prevent
regressions like https://github.com/ClangBuiltLinux/linux/issues/634.

This is disabled for kernel/common because the kernel does not shut down
cleanly; I don't think this is worth exploring because Android does not
care about EFI as far as I am aware.

I considered checking in the fd files so that other distributions could
take advantage of this but these files are over 50MB, which is too much
of a burden to force on everyone.

Presubmit: https://travis-ci.com/nathanchance/continuous-integration/builds/123450697

[skip ci]

Suggested-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>