images/ppc64le: Bump skiboot to 6.5.1
There is currently an issue with booting a ppc64le kernel with QEMU
4.2.0-rc4 with the skiboot we package in this repo (done in #138), see
the link in #221 below.

A bisect points to the QEMU commit linked below. Updating to the latest
stable skiboot 6.5.1 fixes the issue and does not break QEMU 3.1.0 that
we currently use for our Docker image.

[skip ci]

Link: https://github.com/ClangBuiltLinux/continuous-integration/pull/221#issuecomment-561983585
Link: https://git.qemu.org/?p=qemu.git;a=commit;h=f30c843ced5055fde71d28d10beb15af97fdfe39
Presubmit: https://travis-ci.com/nathanchance/continuous-integration/builds/139635572
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>