patches: llvm-11: Split kernel-all into version specific folders
Right now, all of the trees need this patch but the patch has been
picked up and will be in linux-next/mainline soon. After that, it will
filter into stable then into Android. That cascade will take some time
so we need the freedom to stop the patch application for some branches
and not others.

Do this by splitting up the kernel-all folder into all of the different
repos we support.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>