driver.sh: Adapt for boot-utils
This script does two things: build a Linux kernel from source and boot
it. Most people have their own build workflows but not many have a boot
workflow. To make it easy for people to boot Linux kernels that they
have built, move the QEMU booting into a new repo, boot-utils, which
allows people to clone just that repo for the booting workflow and
allows us to continue to boot Linux kernels properly.

That script handles all of the QEMU arguments and image locations. We
just have to pass it the build tree and architecture we built.

The image_name variable has been renamed to make_target, since that is
what it is really there for.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>