driver: Change MIPS config to malta_kvm_guest_defconfig
malta_defconfig does not have CONFIG_BLK_DEV_INITRD enabled, which needs
to be enabled for boot-qemu.sh to work properly. driver.sh enables that
as part of the extra configs that we enable for test coverage (see
common.config) but that confuses people when they go to try and build a
MIPS image and boot it on their own.

Change the config to malta_kvm_guest_defconfig, which has the same
drivers enabled but we get a few extra configs that make booting in QEMU
more proper so we do not lose test coverage and we can still point
people here to show what configs we are using and expect to work.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>