travis: Move MIPS big endian to ld.bfd on -next for now
The series to fix https://github.com/ClangBuiltLinux/linux/issues/785
exposed an issue with ld.lld, which breaks the build.

Move this target to ld.bfd while we discuss how to fix it so that this
target is not blocked for testing.

[skip ci]

Link: https://github.com/ClangBuiltLinux/linux/issues/1025
Presubmit: https://travis-ci.com/github/nathanchance/continuous-integration/builds/166566747
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>