travis: Disable ppc32 for now
Our CI will be perpetually red with this enabled while the bug linked in
the comments is being solved. Unwinding the patches that introduced this
failure will be rather difficult as they came as part of a series with a
merge (from what I can tell) and maintaining that is not worth the
hassle.

While we're here, remove the linux cron comment; mainline does not have
any crons.

[skip ci]

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>