driver.sh: Disable torture tests for PowerPC
The torture tests continue to flake with QEMU 4.2.0 even though I
thought they would not. This has been fixed with QEMU 5.0.0 but unless
we build it ourselves (and I am not really interested in doing so),
there is no fix for this aside from disabling the tests.

Do this so we continue to avoid false positives and can spot real issues
easily when they happen.

Link: https://github.com/ClangBuiltLinux/continuous-integration/issues/262
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>