travis: Disable RISC-V boot again
After fixing the build errors, we still cannot boot. This appears to be
a bug with Ubuntu/Debian's packaging of QEMU, as stock 4.2.0 does not
appear to have the issue, nor does 5.0.0. 5.1.0 has an unrelated
regression.

The solution is most likely to build QEMU ourselves, as we have run into
bugs that were fixed in a newer release before. Further TODO/exploration
as a follow up to this patch.

[skip ci]

Link: https://travis-ci.com/github/nathanchance/continuous-integration/builds/181735734
Presubmit: https://travis-ci.com/github/nathanchance/continuous-integration/builds/181740252
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>