driver: Increase timeout for arm64 to 8m
arm64 builds have recently been timing out during boot on mainline and
next, regardless of LLVM version. It is possible this is something that
should be investigated on the kernel side but local reproduction is a
bit more difficult due to higher power cores than what are available on
Travis. Just increase the timeout to 8m to be sure that we are hung.

[skip ci]

Presubmit: https://travis-ci.com/github/nathanchance/continuous-integration/builds/183784691
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>